chore: pki util
