+++ pkg/chaosctl/physicalmachine/generate.go
@@ -17,22 +17,12 @@ package physicalmachine
 
 import (
 	"crypto"
-	"crypto/ecdsa"
-	"crypto/elliptic"
-	cryptorand "crypto/rand"
-	"crypto/rsa"
 	"crypto/x509"
-	"crypto/x509/pkix"
-	"fmt"
-	"math"
-	"math/big"
-	"time"
+	"io/ioutil"
 
 	"github.com/go-logr/logr"
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
-	certutil "k8s.io/client-go/util/cert"
-	"k8s.io/client-go/util/keyutil"
 )
 
 type PhysicalMachineGenerateOptions struct {
@@ -54,7 +44,7 @@ func NewPhysicalMachineGenerateCmd(logger logr.Logger) (*cobra.Command, error) {
 		SilenceErrors: true,
 		SilenceUsage:  true,
 		RunE: func(cmd *cobra.Command, args []string) error {
-			return nil
+			return generateOption.Run()
 		},
 	}
 	generateCmd.PersistentFlags().StringVar(&generateOption.outputPath, "path", "/etc/chaosd/ssl", "path to save generated certs")
@@ -63,112 +53,29 @@ func NewPhysicalMachineGenerateCmd(logger logr.Logger) (*cobra.Command, error) {
 	return generateCmd, nil
 }
 
+func (o *PhysicalMachineGenerateOptions) Run() error {
+	caCert, caKey, err := GetChaosdCAFileFromFile(o.caCertFile, o.caKeyFile)
+	if err != nil {
+		return err
+	}
+
+	serverCert, serverKey, err := NewCertAndKey(caCert, caKey)
+	if err != nil {
+		return err
+	}
+
+	return WriteCertAndKey(o.outputPath, "chaosd", serverCert, serverKey)
+}
+
 func GetChaosdCAFileFromFile(caCertFile, caKeyFile string) (*x509.Certificate, crypto.Signer, error) {
-
-	return nil, nil, nil
-}
-
-func GenerateChaosdServerCerts() (*x509.Certificate, crypto.Signer, error) {
-
-	return nil, nil, nil
-}
-
-const (
-	// PrivateKeyBlockType is a possible value for pem.Block.Type.
-	PrivateKeyBlockType = "PRIVATE KEY"
-	// PublicKeyBlockType is a possible value for pem.Block.Type.
-	PublicKeyBlockType = "PUBLIC KEY"
-	// CertificateBlockType is a possible value for pem.Block.Type.
-	CertificateBlockType = "CERTIFICATE"
-	// RSAPrivateKeyBlockType is a possible value for pem.Block.Type.
-	RSAPrivateKeyBlockType = "RSA PRIVATE KEY"
-	rsaKeySize             = 2048
-	// CertificateValidity defines the validity for all the signed certificates generated by kubeadm
-	CertificateValidity = time.Hour * 24 * 365
-)
-
-func ParsePrivateKey(data []byte) (crypto.Signer, error) {
-	privKey, err := keyutil.ParsePrivateKeyPEM(data)
+	certData, err := ioutil.ReadFile(caCertFile)
 	if err != nil {
-		return nil, fmt.Errorf("error reading private key file: %v", err)
+		return nil, nil, errors.Wrap(err, "cannot read cert file")
 	}
 
-	var key crypto.Signer
-	// Allow RSA and ECDSA formats only
-	switch k := privKey.(type) {
-	case *rsa.PrivateKey:
-		key = k
-	case *ecdsa.PrivateKey:
-		key = k
-	default:
-		return nil, errors.New("the private key file is neither in RSA nor ECDSA format")
-	}
-	return key, nil
-}
-
-func ParseCert(data []byte) (*x509.Certificate, error) {
-	caCerts, err := certutil.ParseCertsPEM(data)
+	keyData, err := ioutil.ReadFile(caKeyFile)
 	if err != nil {
-		return nil, errors.Wrap(err, "parse certs pem failed")
+		return nil, nil, errors.Wrap(err, "cannot read private key file")
 	}
-	return caCerts[0], nil
-}
-
-// NewCertAndKey creates new certificate and key by passing the certificate authority certificate and key
-func NewCertAndKey(caCert *x509.Certificate, caKey crypto.Signer) (*x509.Certificate, crypto.Signer, error) {
-	key, err := NewPrivateKey(x509.RSA)
-	if err != nil {
-		return nil, nil, errors.Wrap(err, "unable to create private key")
-	}
-
-	cert, err := NewSignedCert(key, caCert, caKey, false)
-	if err != nil {
-		return nil, nil, errors.Wrap(err, "unable to sign certificate")
-	}
-
-	return cert, key, nil
-}
-
-func NewPrivateKey(keyType x509.PublicKeyAlgorithm) (crypto.Signer, error) {
-	if keyType == x509.ECDSA {
-		return ecdsa.GenerateKey(elliptic.P256(), cryptorand.Reader)
-	}
-
-	return rsa.GenerateKey(cryptorand.Reader, rsaKeySize)
-}
-
-// NewSignedCert creates a signed certificate using the given CA certificate and key
-func NewSignedCert(key crypto.Signer, caCert *x509.Certificate, caKey crypto.Signer, isCA bool) (*x509.Certificate, error) {
-	serial, err := cryptorand.Int(cryptorand.Reader, new(big.Int).SetInt64(math.MaxInt64))
-	if err != nil {
-		return nil, err
-	}
-
-	keyUsage := x509.KeyUsageKeyEncipherment | x509.KeyUsageDigitalSignature
-	if isCA {
-		keyUsage |= x509.KeyUsageCertSign
-	}
-
-	notAfter := time.Now().Add(CertificateValidity).UTC()
-
-	certTmpl := x509.Certificate{
-		Subject: pkix.Name{
-			CommonName:   "",
-			Organization: []string{""},
-		},
-		DNSNames: []string{""},
-		// IPAddresses:           []net.IP{},
-		SerialNumber: serial,
-		NotBefore:    caCert.NotBefore,
-		NotAfter:     notAfter,
-		KeyUsage:     keyUsage,
-		// ExtKeyUsage:           cfg.Usages,
-		BasicConstraintsValid: true,
-		IsCA:                  isCA,
-	}
-	certDERBytes, err := x509.CreateCertificate(cryptorand.Reader, &certTmpl, caCert, key.Public(), caKey)
-	if err != nil {
-		return nil, err
-	}
-	return x509.ParseCertificate(certDERBytes)
+	return ParseCertAndKey(certData, keyData)
 }
